#8 : add reporter interface
