+++ exceptions/NotImplementedException.js
@@ -1,0 +1,6 @@
+export class NotImplementedException extends Error {
+  constructor(message) {
+    super(message);
+    this.name = "NotImplementedException";
+  }
+}
